refactor parameter / version update tests #151
