+++ sensor/tests/over-the-air-update/test_local_libraries.py
@@ -1,0 +1,9 @@
+import pytest
+
+
+@pytest.mark.over_the_air_update
+def test_local_libraries() -> None:
+    """checks whether the raspberry pi specific libraries can be loaded"""
+
+    import board
+    import RPi.GPIO
