+++ sensor/tests/over-the-air-update/test_new_config.py
@@ -1,19 +1,15 @@
 import json
 import os
 import pytest
-import os
 from os.path import dirname
-import pytest
-import dotenv
 from src import custom_types
 
 PROJECT_DIR = dirname(dirname(dirname(os.path.abspath(__file__))))
 CONFIG_PATH = os.path.join(PROJECT_DIR, "config", "config.json")
 
 
-@pytest.mark.parameter_update
-@pytest.mark.version_update
-def test_local_config() -> None:
+@pytest.mark.over_the_air_update
+def test_new_config() -> None:
     """checks whether the local config/config.json makes sense:"""
 
     # verify that config matches template types
@@ -37,29 +33,3 @@ def test_local_config() -> None:
     assert len(calibration_gas_bottle_ids) == len(
         unique_calibration_gas_concentrations
     ), "multiple calibration cylinders use the same bottle id"
-
-
-@pytest.mark.version_update
-def test_local_env_vars() -> None:
-    """checks whether the local config/.env can be loaded"""
-
-    env_file_path = os.path.join(PROJECT_DIR, "config", ".env")
-    assert os.path.isfile(env_file_path), f"{env_file_path} is not a file"
-
-    dotenv.load_dotenv(env_file_path)
-    custom_types.MQTTConfig(
-        station_identifier=os.environ.get("HERMES_MQTT_IDENTIFIER"),
-        mqtt_url=os.environ.get("HERMES_MQTT_URL"),
-        mqtt_port=os.environ.get("HERMES_MQTT_PORT"),
-        mqtt_username=os.environ.get("HERMES_MQTT_USERNAME"),
-        mqtt_password=os.environ.get("HERMES_MQTT_PASSWORD"),
-        mqtt_base_topic=os.environ.get("HERMES_MQTT_BASE_TOPIC"),
-    )
-
-
-@pytest.mark.version_update
-def test_local_libraries() -> None:
-    """checks whether the raspberry pi specific libraries can be loaded"""
-
-    import board
-    import RPi.GPIO
